Add input control and can be compiled
The file has been modified, now it can be compiled (with warnings but no
errors).
Now it receives a input file (the test files i included) where it gets
the data from.

also check the MAX_THREADS variable, that can be changed depending on
the Compute Capability of ur own graphics card. We are trying with CUDA
v4 or v5, so it won't be a problem with those values.

I also formatted the variables to be the same as the README file in the
repo. So please check the README to understand or remember the Neighbor
Joining Algorithm.